backend/requirements.txt
Flask
python-dotenv
SQLAlchemy
psycopg2-binary
requests
gunicorn
google-auth
google-auth-oauthlib
google-api-python-client
